# docs/api/openAPI.yaml
...
openapi: 3.1.0
info:
  title: Industry Core Hub Backend API
  version: 0.0.1
paths:
  '/part-management/catalog-part/{manufacturer_id}/manufacturerPartId':
    get:
      tags:
        - Part Management
      summary: Part Management Get Catalog Part
      operationId: >-
        part_management_get_catalog_part_part_management_catalog_part__manufacturer_id___manufacturer_part_id__get
      parameters:
        - name: manufacturer_id
          in: path
          required: true
          schema:
            type: string
            title: Manufacturer Id
        - name: manufacturer_part_id
          in: path
          required: true
          schema:
            type: string
            title: Manufacturer Part Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPartReadWithStatus'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /part-management/catalog-part:
    get:
      tags:
        - Part Management
      summary: Part Management Get Catalog Parts
      operationId: part_management_get_catalog_parts_part_management_catalog_part_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SimpleCatalogPartReadWithStatus'
                type: array
                title: >-
                  Response Part Management Get Catalog Parts Part Management
                  Catalog Part Get
    post:
      tags:
        - Part Management
      summary: Part Management Create Catalog Part
      operationId: part_management_create_catalog_part_part_management_catalog_part_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogPartCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPartReadWithStatus'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BusinessPartnerRead:
      properties:
        name:
          type: string
          title: Name
          description: The unique name of the business partner.
        bpnl:
          type: string
          title: Bpnl
          description: The Catena-X Business Partner Number (BPNL) of the business partner.
      type: object
      required:
        - name
        - bpnl
      title: BusinessPartnerRead
      description: Represents a partner with a validated name field.
    CatalogPartCreate:
      properties:
        manufacturerId:
          type: string
          title: Manufacturerid
          description: The BPNL (manufactuer ID) of the part to register.
        manufacturerPartId:
          type: string
          title: Manufacturerpartid
          description: The manufacturer part ID of the part.
        name:
          type: string
          title: Name
          description: The name of the part.
        category:
          anyOf:
            - type: string
            - type: 'null'
          title: Category
          description: The category of the part.
        bpns:
          anyOf:
            - type: string
            - type: 'null'
          title: Bpns
          description: The site number (BPNS) the part is attached to.
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
          description: The decription of the part.
        materials:
          items:
            $ref: '#/components/schemas/Material'
          type: array
          title: Materials
          description: 'List of materials, e.g. [{''name'':''aluminum'',''share'':''20''}]'
          default: []
        width:
          anyOf:
            - $ref: '#/components/schemas/Measurement'
            - type: 'null'
          description: The width of the part.
        height:
          anyOf:
            - $ref: '#/components/schemas/Measurement'
            - type: 'null'
          description: The height of the part.
        length:
          anyOf:
            - $ref: '#/components/schemas/Measurement'
            - type: 'null'
          description: The length of the part.
        weight:
          anyOf:
            - $ref: '#/components/schemas/Measurement'
            - type: 'null'
          description: The weight of the part.
        customerPartIds:
          anyOf:
            - additionalProperties:
                $ref: '#/components/schemas/BusinessPartnerRead'
              type: object
            - type: 'null'
          title: Customerpartids
          description: >-
            The list of customer part IDs mapped to the respective Business
            Partners.
          default: {}
      type: object
      required:
        - manufacturerId
        - manufacturerPartId
        - name
      title: CatalogPartCreate
    CatalogPartReadWithStatus:
      properties:
        manufacturerId:
          type: string
          title: Manufacturerid
          description: The BPNL (manufactuer ID) of the part to register.
        manufacturerPartId:
          type: string
          title: Manufacturerpartid
          description: The manufacturer part ID of the part.
        name:
          type: string
          title: Name
          description: The name of the part.
        category:
          anyOf:
            - type: string
            - type: 'null'
          title: Category
          description: The category of the part.
        bpns:
          anyOf:
            - type: string
            - type: 'null'
          title: Bpns
          description: The site number (BPNS) the part is attached to.
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
          description: The decription of the part.
        materials:
          items:
            $ref: '#/components/schemas/Material'
          type: array
          title: Materials
          description: 'List of materials, e.g. [{''name'':''aluminum'',''share'':''20''}]'
          default: []
        width:
          anyOf:
            - $ref: '#/components/schemas/Measurement'
            - type: 'null'
          description: The width of the part.
        height:
          anyOf:
            - $ref: '#/components/schemas/Measurement'
            - type: 'null'
          description: The height of the part.
        length:
          anyOf:
            - $ref: '#/components/schemas/Measurement'
            - type: 'null'
          description: The length of the part.
        weight:
          anyOf:
            - $ref: '#/components/schemas/Measurement'
            - type: 'null'
          description: The weight of the part.
        customerPartIds:
          anyOf:
            - additionalProperties:
                $ref: '#/components/schemas/BusinessPartnerRead'
              type: object
            - type: 'null'
          title: Customerpartids
          description: >-
            The list of customer part IDs mapped to the respective Business
            Partners.
          default: {}
        status:
          $ref: '#/components/schemas/SharingStatus'
          description: >-
            The status of the part. (0: draft, 1:pending, 2: registered, 3:
            shared)
      type: object
      required:
        - manufacturerId
        - manufacturerPartId
        - name
        - status
      title: CatalogPartReadWithStatus
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Material:
      properties:
        name:
          type: string
          title: Name
          description: Name of the material
        share:
          type: integer
          title: Share
          description: Share of the material in percent. 0-100
      type: object
      required:
        - name
        - share
      title: Material
    Measurement:
      properties:
        value:
          type: number
          title: Value
          description: Numeric value of the measurement
        unit:
          $ref: '#/components/schemas/Unit'
          description: Unit of measure
      type: object
      required:
        - value
        - unit
      title: Measurement
...
    SharingStatus:
      type: integer
      enum:
        - 0
        - 1
        - 2
        - 3
      title: SharingStatus
      description: 'The status of the part. (0: draft, 1: pending, 2: registered, 3: shared)'
    SimpleCatalogPartReadWithStatus:
      properties:
        manufacturerId:
          type: string
          title: Manufacturerid
          description: The BPNL (manufactuer ID) of the part to register.
        manufacturerPartId:
          type: string
          title: Manufacturerpartid
          description: The manufacturer part ID of the part.
        name:
          type: string
          title: Name
          description: The name of the part.
        category:
          anyOf:
            - type: string
            - type: 'null'
          title: Category
          description: The category of the part.
        bpns:
          anyOf:
            - type: string
            - type: 'null'
          title: Bpns
          description: The site number (BPNS) the part is attached to.
        status:
          $ref: '#/components/schemas/SharingStatus'
          description: >-
            The status of the part. (0: draft, 1:pending, 2: registered, 3:
            shared)
      type: object
      required:
        - manufacturerId
        - manufacturerPartId
        - name
        - status
      title: SimpleCatalogPartReadWithStatus
...
    Unit:
      type: string
      enum:
        - mm
        - cm
        - m
        - g
        - kg
      title: Unit
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...